refactor(slots): extract container-image pull orchestration to slots/image_pull
P3-routers §J: move the slot container-image pull vertical out of
routes/slots.py into a new deep module hal0.slots.image_pull — the
layer-oriented ImagePullJob, the background pull runner, and the
profile→image resolver + present|missing inspection that had been copied
inline across the POST / stream / status handlers. The pull_slot_image*
handlers are now request→service→envelope shells (the SSE + BackgroundTasks
wrappers stay in the route). Route paths/methods/payloads/status unchanged.

Back-compat: routes/slots keeps _ImagePullJob + _run_image_pull re-export
bindings so the test-suite's import + patch seams keep working.

slots.py: 1629 → 1513 lines.
